+++ src/main/java/com/CarRentProject/service/CarServiceImpl.java
@@ -26,7 +26,7 @@ public class CarServiceImpl implements CarService {
     }
 
     @Override
-    public Car getCarById(Integer id) {
+    public Car getCarById(Long id) {
         return carRepository.findById(id);
     }
